+++ src/StoreApi/Exceptions/RouteException.php
@@ -1,8 +1,8 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Routes;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Exceptions;
 
 /**
- * ReserveStockRouteExceptionException class.
+ * RouteException class.
  *
  * @internal This API is used internally by Blocks--it is still in flux and may be subject to revisions.
  */
