+++ src/StoreApi/Routes/V1/Batch.php
@@ -2,6 +2,7 @@
 namespace Automattic\WooCommerce\Blocks\StoreApi\Routes\V1;
 
 use Automattic\WooCommerce\Blocks\StoreApi\Routes\RouteInterface;
+use Automattic\WooCommerce\Blocks\StoreApi\Exceptions\RouteException;
 use WP_REST_Request;
 use WP_REST_Response;
