+++ src/StoreApi/Routes/V1/Cart.php
@@ -38,7 +38,6 @@ class Cart extends AbstractCartRoute {
 	/**
 	 * Handle the request and return a valid response for this endpoint.
 	 *
-	 * @throws RouteException On error.
 	 * @param \WP_REST_Request $request Request object.
 	 * @return \WP_REST_Response
 	 */
