+++ src/StoreApi/Routes/V1/CartCoupons.php
@@ -1,6 +1,8 @@
 <?php
 namespace Automattic\WooCommerce\Blocks\StoreApi\Routes\V1;
 
+use Automattic\WooCommerce\Blocks\StoreApi\Exceptions\RouteException;
+
 /**
  * CartCoupons class.
  *
