+++ src/StoreApi/Routes/V1/CartUpdateItem.php
@@ -45,8 +45,8 @@ class CartUpdateItem extends AbstractCartRoute {
 
 	/**
 	 * Handle the request and return a valid response for this endpoint.
+	 * .
 	 *
-	 * @throws RouteException On error.
 	 * @param \WP_REST_Request $request Request object.
 	 * @return \WP_REST_Response
 	 */
