+++ src/StoreApi/Routes/V1/ProductCollectionData.php
@@ -41,7 +41,6 @@ class ProductCollectionData extends AbstractRoute {
 	/**
 	 * Get a collection of posts and add the post title filter option to \WP_Query.
 	 *
-	 * @throws RouteException On error.
 	 * @param \WP_REST_Request $request Request object.
 	 * @return \WP_REST_Response
 	 */
