+++ tests/php/Domain/Services/TestExtendRestApi.php
@@ -9,7 +9,6 @@ use Automattic\WooCommerce\Blocks\StoreApi\Formatters;
 use Automattic\WooCommerce\Blocks\StoreApi\Formatters\CurrencyFormatter;
 use Automattic\WooCommerce\Blocks\StoreApi\Formatters\HtmlFormatter;
 use Automattic\WooCommerce\Blocks\StoreApi\Formatters\MoneyFormatter;
-use Automattic\WooCommerce\Blocks\StoreApi\Routes\RouteException;
 use Exception;
 use Automattic\WooCommerce\Blocks\Domain\Services\FeatureGating;
 use Automattic\WooCommerce\Blocks\Domain\Package as DomainPackage;
@@ -47,10 +46,13 @@ class TestExtendRestApi extends TestCase {
 			return null;
 		};
 		// set listening for exceptions
-		add_action( 'woocommerce_caught_exception', function($exception_object){
-			$this->caught_exception = $exception_object;
-			throw $exception_object;
-		});
+		add_action(
+			'woocommerce_caught_exception',
+			function( $exception_object ) {
+				$this->caught_exception = $exception_object;
+				throw $exception_object;
+			}
+		);
 	}
 
 	/**
@@ -58,10 +60,10 @@ class TestExtendRestApi extends TestCase {
 	 */
 	public function test_register_callback() {
 		$this->mock_extend->register_update_callback(
-			[
+			array(
 				'namespace' => 'test-plugin',
 				'callback'  => $this->dummy,
-			]
+			)
 		);
 		$this->assertSame( $this->dummy, $this->mock_extend->get_update_callback( 'test-plugin' ) );
 	}
@@ -71,11 +73,11 @@ class TestExtendRestApi extends TestCase {
 	 */
 	public function test_fail_register_callback() {
 		$this->expectException( Exception::class );
-		$this->expectExceptionMessage('You must provide a plugin namespace when extending a Store REST endpoint.');
+		$this->expectExceptionMessage( 'You must provide a plugin namespace when extending a Store REST endpoint.' );
 		$this->mock_extend->register_update_callback(
-			[
+			array(
 				'callback' => $this->dummy,
-			]
+			)
 		);
 	}
 
@@ -84,12 +86,12 @@ class TestExtendRestApi extends TestCase {
 	 */
 	public function test_fail_get_callback() {
 		$this->expectException( Exception::class );
-		$this->expectExceptionMessage('There is no such namespace registered: nonexistent-plugin.');
+		$this->expectExceptionMessage( 'There is no such namespace registered: nonexistent-plugin.' );
 		$this->mock_extend->register_update_callback(
-			[
+			array(
 				'namespace' => 'test-plugin',
 				'callback'  => $this->dummy,
-			]
+			)
 		);
 		$this->mock_extend->get_update_callback( 'nonexistent-plugin' );
 	}
@@ -99,11 +101,11 @@ class TestExtendRestApi extends TestCase {
 	 */
 	public function test_fail_get_callback_with_uncallable() {
 		$this->expectException( Exception::class );
-		$this->expectExceptionMessage('There is no valid callback supplied to register_update_callback.');
+		$this->expectExceptionMessage( 'There is no valid callback supplied to register_update_callback.' );
 		$this->mock_extend->register_update_callback(
-			[
+			array(
 				'namespace' => 'test-plugin',
-			]
+			)
 		);
 		$this->mock_extend->get_update_callback( 'nonexistent-plugin' );
 	}
